fix(collector): allow cross-compilation on non-Linux platforms
 factory.go unconditionally imported collector/config, which pulled in
  tracer → rlimit → unix.Prlimit (Linux-only), breaking Darwin builds.

Merge factory.go into factory_linux.go so collector/config is only
imported under the linux build constraint. Add a standalone NewFactory
stub to factory_others.go that avoids the problematic import chain.
