--- collector/factory.go
@@ -1,45 +1,0 @@
-// Copyright The OpenTelemetry Authors
-// SPDX-License-Identifier: Apache-2.0
-
-//go:generate go tool -modfile=../internal/tools/go.mod mdatagen metadata.yaml
-
-package collector // import "go.opentelemetry.io/ebpf-profiler/collector"
-
-import (
-	"errors"
-	"time"
-
-	"go.opentelemetry.io/collector/component"
-	"go.opentelemetry.io/collector/receiver"
-	"go.opentelemetry.io/collector/receiver/xreceiver"
-
-	"go.opentelemetry.io/ebpf-profiler/collector/config"
-	"go.opentelemetry.io/ebpf-profiler/collector/internal/metadata"
-)
-
-var errInvalidConfig = errors.New("invalid config")
-
-// NewFactory creates a factory for the receiver.
-func NewFactory() receiver.Factory {
-	return xreceiver.NewFactory(
-		metadata.Type,
-		defaultConfig,
-		xreceiver.WithProfiles(BuildProfilesReceiver(), metadata.ProfilesStability))
-}
-
-func defaultConfig() component.Config {
-	return &config.Config{
-		ReporterInterval:       5 * time.Second,
-		ReporterJitter:         0.2,
-		MonitorInterval:        5 * time.Second,
-		SamplesPerSecond:       20,
-		ProbabilisticInterval:  1 * time.Minute,
-		ProbabilisticThreshold: 100,
-		Tracers:                "all",
-		ClockSyncInterval:      3 * time.Minute,
-		MaxGRPCRetries:         5,
-		MaxRPCMsgSize:          32 << 20, // 32 MiB,
-		BPFFSRoot:              "/sys/fs/bpf/",
-		ErrorMode:              config.PropagateError,
-	}
-}
