+++ collector/factory_linux_test.go
@@ -1,0 +1,46 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+//go:build linux && (amd64 || arm64)
+
+package collector
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/require"
+	"go.opentelemetry.io/collector/component"
+	"go.opentelemetry.io/collector/consumer/consumertest"
+	"go.opentelemetry.io/collector/receiver/receivertest"
+)
+
+func TestCreateProfilesReceiver(t *testing.T) {
+	for _, tt := range []struct {
+		name   string
+		config component.Config
+
+		wantError error
+	}{
+		{
+			name:   "Default config",
+			config: NewFactory().CreateDefaultConfig(),
+		},
+		{
+			name:      "Nil config",
+			wantError: errInvalidConfig,
+		},
+	} {
+		t.Run(tt.name, func(t *testing.T) {
+			t.Parallel()
+			typ, err := component.NewType("ProfilesReceiver")
+			require.NoError(t, err)
+			_, err = BuildProfilesReceiver()(
+				t.Context(),
+				receivertest.NewNopSettings(typ),
+				tt.config,
+				consumertest.NewNop(),
+			)
+			require.ErrorIs(t, err, tt.wantError)
+		})
+	}
+}
