+++ collector/factory_others.go
@@ -13,7 +13,17 @@ import (
 	"go.opentelemetry.io/collector/consumer/xconsumer"
 	"go.opentelemetry.io/collector/receiver"
 	"go.opentelemetry.io/collector/receiver/xreceiver"
+
+	"go.opentelemetry.io/ebpf-profiler/collector/internal/metadata"
 )
+
+// NewFactory creates a factory for the receiver.
+func NewFactory() receiver.Factory {
+	return xreceiver.NewFactory(
+		metadata.Type,
+		func() component.Config { return &struct{}{} },
+		xreceiver.WithProfiles(BuildProfilesReceiver(), metadata.ProfilesStability))
+}
 
 func BuildProfilesReceiver(options ...Option) xreceiver.CreateProfilesFunc {
 	return func(_ context.Context,
@@ -22,6 +32,5 @@ func BuildProfilesReceiver(options ...Option) xreceiver.CreateProfilesFunc {
 		_ xconsumer.Profiles,
 	) (xreceiver.Profiles, error) {
 		return nil, errors.New("profiling receiver is only supported on Linux and arm64 or amd64")
-
 	}
 }
